refactor: add new vscode-ui package (#10470)
* refactor: vscode-ui package

* ci: ui package

* test: ut

* fix: error

* fix: error

* fix: error

* test: ut

* test: ut

* docs: readme
