+++ .github/scripts/lernaDeps.json
@@ -39,5 +39,9 @@
   ],
   "template": ["templates"],
   "extra-shot-mocha": ["packages/extra-shot-mocha"],
-  "eslint-plugin-teamsfx": ["packages/eslint-plugin-teamsfx"]
+  "eslint-plugin-teamsfx": ["packages/eslint-plugin-teamsfx"],
+  "ui": [
+    "packages/api",
+    "packages/vscode-ui"
+  ]
 }
