+++ packages/vscode-extension/test/extension/qm/vsc_ui.test.ts
@@ -30,8 +30,8 @@ import {
   SingleSelectConfig,
   UserError,
 } from "@microsoft/teamsfx-api";
-import { UserCancelError } from "@microsoft/teamsfx-core";
-import { FxQuickPickItem, VsCodeUI } from "../../../src/qm/vsc_ui";
+import { FxQuickPickItem, UserCancelError } from "@microsoft/vscode-ui";
+import { VsCodeUI } from "../../../src/qm/vsc_ui";
 import { ExtTelemetry } from "../../../src/telemetry/extTelemetry";
 import { sleep } from "../../../src/utils/commonUtils";
 import { VsCodeLogProvider } from "../../../src/commonlib/log";
@@ -93,7 +93,7 @@ describe("UI Unit Tests", async () => {
       sinon.stub(window, "createQuickPick").callsFake(() => {
         return mockQuickPick;
       });
-      const telemetryStub = sinon.stub(ExtTelemetry, "sendTelemetryEvent");
+      // const telemetryStub = sinon.stub(ExtTelemetry, "sendTelemetryEvent");
 
       const result = await ui.selectFolder(config);
 
@@ -101,11 +101,11 @@ describe("UI Unit Tests", async () => {
       if (result.isOk()) {
         expect(result.value.result).to.equal("default folder");
       }
-      expect(
-        telemetryStub.calledOnceWith("select-folder", {
-          "selected-option": "default",
-        })
-      ).is.true;
+      // expect(
+      //   telemetryStub.calledOnceWith("select-folder", {
+      //     "selected-option": "default",
+      //   })
+      // ).is.true;
       sinon.restore();
     });
