+++ packages/vscode-ui/.eslintignore
@@ -1,0 +1,2 @@
+templates/plugins/resource/spfx/*
+src/plugins/resource/spfx/authCode.ts
