+++ packages/vscode-ui/.gitignore
@@ -1,0 +1,10 @@
+node_modules
+build
+coverage
+.nyc_output
+.npmrc
+.env
+.prettierrc.json
+.fx
+templates/**/*.zip
+tests/**/TeamsSPFxApp.zip
