+++ packages/vscode-ui/.mocharc.js
@@ -1,0 +1,15 @@
+"use strict";
+
+module.exports = {
+  color: true,
+  delay: false,
+  diff: true,
+  parallel: false,
+  recursive: false,
+  reporter: "spec",
+  require: "ts-node/register",
+  retries: 1,
+  slow: "75",
+  timeout: 0,
+  extensions: ["ts", "tsx"],
+};
