+++ packages/vscode-ui/.prettierignore
@@ -1,0 +1,4 @@
+templates/plugins/resource/spfx
+.prettierignore
+*.bicep
+*.png
