+++ packages/vscode-ui/.prettierrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require("../prettier-config"),
+};
