+++ packages/vscode-ui/package.json
@@ -1,0 +1,69 @@
+{
+  "name": "@microsoft/vscode-ui",
+  "version": "1.0.0",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "license": "MIT",
+  "publisher": "TeamsDevApp",
+  "scripts": {
+    "lint:staged": "lint-staged",
+    "lint": "eslint \"src/**/*.ts\" \"tests/**/*.ts\"",
+    "test:unit": "nyc mocha \"tests/**/*.test.ts\" ",
+    "clean": "rm -rf build",
+    "build": "rimraf build && npx tsc -p ./",
+    "lint:fix": "eslint --fix \"src/**/*.ts\" \"tests/**/*.ts\"",
+    "prepublishOnly": "npm run build",
+    "package": "rimraf build && webpack --mode production --config ./webpack.config.js",
+    "check-sensitive": "npx eslint --plugin 'no-secrets' --cache --ignore-pattern 'package.json' --ignore-pattern 'package-lock.json'",
+    "precommit": "npm run check-sensitive && lint-staged",
+    "installAll": "npm run install:tabs"
+  },
+  "dependencies": {
+    "@microsoft/teamsfx-api": "workspace:*",
+    "async-mutex": "^0.3.1",
+    "lodash": "^4.17.21"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.2.14",
+    "@types/chai-as-promised": "^7.1.3",
+    "@types/chai-spies": "^1.0.3",
+    "@types/lodash": "^4.14.181",
+    "@types/mocha": "^8.0.4",
+    "@types/node": "^14.14.21",
+    "@types/sinon": "^9.0.9",
+    "@types/vscode": "^1.66.0",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "chai-spies": "^1.0.0",
+    "copy-webpack-plugin": "^6.4.1",
+    "eslint": "^7.22.0",
+    "eslint-plugin-header": "^3.1.1",
+    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-no-secrets": "^0.8.9",
+    "eslint-plugin-prettier": "^4.0.0",
+    "lint-staged": "^10.5.4",
+    "mocha": "^10.2.0",
+    "mocked-env": "^1.3.5",
+    "nyc": "^15.1.0",
+    "prettier": "^2.4.1",
+    "sinon": "^9.2.2",
+    "source-map-support": "^0.5.19",
+    "terser-webpack-plugin": "^5.3.9",
+    "tslib": "^2.3.1",
+    "ts-loader": "^8.0.3",
+    "ts-node": "^9.1.1",
+    "typescript": "4.3.5"
+  },
+  "files": [
+    "build/**/*"
+  ],
+  "gitHead": "7d60c0765c0ea8c023a26c10d1c93001c597afbb",
+  "publishConfig": {
+    "access": "public"
+  },
+  "lint-staged": {
+    "*.{js,jsx,css,ts,tsx}": [
+      "npx eslint --cache --fix --quiet"
+    ]
+  }
+}
