+++ packages/vscode-ui/src/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+export * from "./ui";
+export * from "./progressHandler";
+export * from "./localize";
+export * from "./error";
